move server instruments related code to one file
