+++ bentoml/server/instruments.py
@@ -1,8 +1,12 @@
+import multiprocessing
+import os
+import shutil
 from timeit import default_timer
 
 from flask import Request
 
 from bentoml import config
+from bentoml.server.utils import logger
 
 
 class InstrumentMiddleware:
@@ -64,3 +68,32 @@ class InstrumentMiddleware:
             endpoint=endpoint, service_version=self.bento_service.version
         ).track_inprogress():
             return self.app(environ, start_response_wrapper)
+
+
+def setup_prometheus_multiproc_dir(lock: multiprocessing.Lock = None):
+    """
+    Set up prometheus_multiproc_dir for prometheus to work in multiprocess mode,
+    which is required when working with Gunicorn server
+
+    Warning: for this to work, prometheus_client library must be imported after
+    this function is called. It relies on the os.environ['prometheus_multiproc_dir']
+    to properly setup for multiprocess mode
+    """
+    if lock is not None:
+        lock.acquire()
+
+    try:
+        prometheus_multiproc_dir = config('instrument').get('prometheus_multiproc_dir')
+        logger.debug(
+            "Setting up prometheus_multiproc_dir: %s", prometheus_multiproc_dir
+        )
+        # Wipe prometheus metrics directory between runs
+        # https://github.com/prometheus/client_python#multiprocess-mode-gunicorn
+        # Ignore errors so it does not fail when directory does not exist
+        shutil.rmtree(prometheus_multiproc_dir, ignore_errors=True)
+        os.makedirs(prometheus_multiproc_dir, exist_ok=True)
+
+        os.environ['prometheus_multiproc_dir'] = prometheus_multiproc_dir
+    finally:
+        if lock is not None:
+            lock.release()
